fix bug uint

+++ tutorial-contents/Using Neural Nets/Evolution Strategy with Neural Nets.py
@@ -90,7 +90,7 @@ def build_net():
 
 def train(net_shapes, net_params, optimizer, utility, pool):
     # pass seed instead whole noise matrix to parallel will save your time
-    noise_seed = np.random.randint(0, 2 ** 32 - 1, size=N_KID).repeat(2)    # mirrored sampling
+    noise_seed = np.random.randint(0, 2 ** 32 - 1, size=N_KID, dtype=np.uint32).repeat(2)    # mirrored sampling
 
     # distribute training in parallel
     jobs = [pool.apply_async(get_reward, (net_shapes, net_params, env, CONFIG['ep_max_step'], CONFIG['continuous_a'],
